Change some Service functions and get movie information from api

+++ database/src/main/java/com/example/database/repository/ActorRepository.java
@@ -3,5 +3,5 @@ package com.example.database.repository;
 import com.example.database.model.entity.Actor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActorRepository extends CrudRepository<Actor, Long> {
+public interface ActorRepository extends CrudRepository<Actor, Integer> {
 }
